% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConf_alk}
\alias{defConf_alk}
\title{Set up conf}
\usage{
defConf_alk(
  years = NULL,
  maxAge,
  spatioTemporal = 1,
  rwBeta0 = 1,
  cutoff = 100,
  cbound = 200,
  meshSimilar = FALSE,
  zone = NULL
)
}
\arguments{
\item{years}{years to include}

\item{maxAge}{max age}

\item{spatioTemporal}{include spatio-tempora: 0-no 1-yes 2- yes but without correlation structure in time}

\item{cutoff}{mesh spesific, min distance between points}

\item{cbound}{mesh spesific, size of boundary meshSimilar}

\item{meshSimilar}{If true we apply the same mesh as in the package spatitemporalIndices}
}
\value{
Configurations to set up the model
}
\description{
Set up conf
}
\details{
defines the configurations
}
