% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fitALK}
\alias{fitALK}
\title{Set up data}
\usage{
fitALK(data, par, conf)
}
\arguments{
\item{data}{data}

\item{conf}{configurations}
}
\value{
Prameters to be provided to TMB
}
\description{
Set up data
}
\details{
define parameters
}
