% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik.alk}
\alias{logLik.alk}
\title{Log likelihood of stALK object}
\usage{
\method{logLik}{alk}(object, ...)
}
\arguments{
\item{object}{fitted object as returned from the fitALK function}
}
\description{
Log likelihood of stALK object
}
\details{
log likelihood of a ALK model run
}
