% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{partable}
\alias{partable}
\title{Print partabld}
\usage{
partable(run)
}
\arguments{
\item{run}{the fitted object as returned from the fitALK function}
}
\value{
Partable..
}
\description{
Print partabld
}
\details{
Print ...
}
