% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotALK}
\alias{plotALK}
\title{Plot ALK}
\usage{
plotALK(run, year, length = NULL, age = NULL, ...)
}
\arguments{
\item{run}{object of class stALK}

\item{year}{Year of interest}

\item{length}{if != NULL, plot spatial ALK based on length and year}
}
\description{
Plot ALK
}
\details{
Function to plot ALK.
}
