% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.alk}
\alias{print.alk}
\title{Print alk object}
\usage{
\method{print}{alk}(x, ...)
}
\arguments{
\item{x}{the fitted object as returned from the fitALK function}
}
\description{
Print alk object
}
\details{
prints the log-likelihood and the main convergence criteria
}
