% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setUpData_alk}
\alias{setUpData_alk}
\title{Set up data}
\usage{
setUpData_alk(d, conf_alk, conf_l = NULL)
}
\arguments{
\item{df}{raw data}

\item{conf}{configurations}
}
\value{
Data to be provided to TMB
}
\description{
Set up data
}
\details{
Prepare data
}
